test(integration): split the push test into plugin tests
for #145
