+++ test/integration/repository.test.js
@@ -1,0 +1,44 @@
+const path = require('path')
+const fs = require('fs')
+const yaml = require('js-yaml')
+const settings = require('../../lib/settings')
+const { initializeNock, loadInstance, repository, teardownNock } = require('./common')
+
+describe('repository', function () {
+  let probot, githubScope
+
+  beforeEach(() => {
+    githubScope = initializeNock()
+    probot = loadInstance()
+  })
+
+  afterEach(() => {
+    teardownNock(githubScope)
+  })
+
+  it('syncs repo settings', async () => {
+    const pathToConfig = path.resolve(__dirname, '..', 'fixtures', 'repository-config.yml')
+    const configFile = Buffer.from(fs.readFileSync(pathToConfig, 'utf8'))
+    const config = yaml.safeLoad(configFile, 'utf8')
+    const encodedConfig = configFile.toString('base64')
+    githubScope
+      .get(`/repos/${repository.owner.name}/${repository.name}/contents/${settings.FILE_NAME}`)
+      .reply(200, { content: encodedConfig, name: 'settings.yml', type: 'file' })
+    githubScope
+      .patch(`/repos/${repository.owner.name}/${repository.name}`, body => {
+        expect(body).toMatchObject(config.repository)
+        return true
+      })
+      .matchHeader('accept', ['application/vnd.github.baptiste-preview+json'])
+      .reply(200)
+
+    await probot.receive({
+      name: 'push',
+      payload: {
+        ref: 'refs/heads/master',
+        repository,
+        commits: [{ modified: [settings.FILE_NAME], added: [] }]
+      }
+    })
+  })
+})
